% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_class.R, R/graph_methods.R
\docType{class}
\name{Graph}
\alias{Graph}
\alias{Graph-class}
\alias{initialize,Graph-method}
\title{Graph Class}
\usage{
\S4method{initialize}{Graph}(.Object, data, coords, dict, directed = TRUE, crs = NA_character_)
}
\description{
This class represents a graph with nodes, edges, and costs.

Graph Constructor Method
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame containing the graph edges and costs.}

\item{\code{coords}}{A data.frame containing the coordinates for each node (optional).}

\item{\code{dict}}{A data.frame mapping the node reference to its integer ID.}
}}

\keyword{classes}
