% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{makegraph}
\alias{makegraph}
\title{Make a Graph}
\usage{
makegraph(from, to, cost, directed = TRUE, coords = NULL, crs = NA_character_)
}
\arguments{
\item{from}{Vector of node IDs representing the start of an edge.}

\item{to}{Vector of node IDs representing the end of an edge.}

\item{cost}{Vector of edge costs.}

\item{directed}{Logical indicating whether the graph is directed or not.}

\item{coords}{An optional data.frame containing the node coordinates.}

\item{crs}{A character string specifying the coordinate reference system (CRS) of the coordinates}
}
\value{
A Graph object.
}
\description{
Create a graph object from edge data.
}
\examples{
\dontrun{
# Create a graph
from  <- c(1, 1, 1, 2, 2, 3, 3, 4, 5, 6, 6, 7)
to    <- c(2, 3, 4, 3, 5, 4, 6, 6, 7, 5, 7, 8)
cost  <- c(5.6, 3.8, 2.0, 1.2, 6.1, 4.2, 2.5, 7.8, 3.3, 1.7, 6.9, 2.2)

graph <- makegraph(from, to, cost)

coords <- data.frame(
id = graph@dict$ref,
lon = c(-73.987, -73.988, -73.986, -73.989, -73.984, -73.985, -73.990, -73.983),
lat = c(40.753, 40.754, 40.755, 40.756, 40.757, 40.752, 40.758, 40.751)
)
graph2 <- makegraph(from, to, cost, coords = coords, crs = "epsg:4326")

}
}
