% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{Graph}
\alias{Graph}
\title{Graph Class}
\description{
This class represents a graph with nodes and edges.
}
\section{Usage}{

\preformatted{
graph <- Graph$new(edge_from, edge_to, edge_cost, edge_dist, node_name, node_x, node_y, crs)
}
}

\section{Methods}{

\describe{
\item{edges()}{Returns a list of edges in the graph.}
\item{nodes()}{Returns a list of nodes in the graph.}
\item{node_dict()}{Returns a named list of node indices in the graph.}
\item{crs()}{Returns the CRS string of the graph.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pointer}}{Pointer to the Graph object in C++.
Initialize Graph Object

Creates a new Graph object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Graph-new}{\code{Graph$new()}}
\item \href{#method-Graph-edges}{\code{Graph$edges()}}
\item \href{#method-Graph-nodes}{\code{Graph$nodes()}}
\item \href{#method-Graph-node_dict}{\code{Graph$node_dict()}}
\item \href{#method-Graph-crs}{\code{Graph$crs()}}
\item \href{#method-Graph-profile}{\code{Graph$profile()}}
\item \href{#method-Graph-activate_profile}{\code{Graph$activate_profile()}}
\item \href{#method-Graph-print}{\code{Graph$print()}}
\item \href{#method-Graph-clone}{\code{Graph$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-new"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$new(
  edge_from,
  edge_to,
  edge_speed,
  edge_length,
  edge_oneway,
  node_name,
  node_x,
  node_y,
  crs
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{edge_from}}{character vector of edge sources.}

\item{\code{edge_to}}{character vector of edge targets.}

\item{\code{edge_speed}}{numeric vector of edge speeds.}

\item{\code{edge_length}}{numeric vector of edge lengths.}

\item{\code{edge_oneway}}{character vector of edge oneway types (one-way: from-to = "FT", one-way: to-from = "TF", two-way = "B", or restricted = "N").}

\item{\code{node_name}}{character vector of node names.}

\item{\code{node_x}}{numeric vector of node x-coordinates.}

\item{\code{node_y}}{numeric vector of node y-coordinates.}

\item{\code{crs}}{character string of the CRS (coordinate reference system).
Get Edges

Returns a list of edges in the graph.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-edges"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-edges}{}}}
\subsection{Method \code{edges()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$edges()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of edges with columns "from", "to", "cost", "speed", "length", and "oneway".
Get Nodes

Returns a list of nodes in the graph.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-nodes"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-nodes}{}}}
\subsection{Method \code{nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$nodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of nodes with columns "id", "X", and "Y".
Get Node Dictionary

Returns a named list of node indices in the graph.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-node_dict"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-node_dict}{}}}
\subsection{Method \code{node_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$node_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list of node indices, with node names as the names and node indices as the values.
Get CRS

Returns the CRS (coordinate reference system) string of the graph.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-crs"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-crs}{}}}
\subsection{Method \code{crs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$crs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string of the CRS.
Get CRS

Returns the active routing profile string of the graph.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-profile"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-profile}{}}}
\subsection{Method \code{profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$profile()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string of the CRS.
Activate Routing Profile

Activates a routing profile for the graph.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-activate_profile"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-activate_profile}{}}}
\subsection{Method \code{activate_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$activate_profile(profile = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{profile}}{A character string specifying the routing profile ("default" = 0, "foot" = 1, "bicycle" = 2, "car" = 3).
Print Graph Summary

Prints a summary of the graph object, including the number of nodes and edges.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-print"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Print
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graph-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Graph-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graph$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
