% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{makegraph}
\alias{makegraph}
\title{Create a Graph object}
\usage{
makegraph(edges, nodes, crs, directed = TRUE)
}
\arguments{
\item{edges}{data.frame with columns "from", "to", "speed" [km/h], "length" [m], "oneway" (one-way: from-to = "FT", one-way: to-from = "TF", two-way = "B", restricted = "N", or pedestiran only = "foot_only" (bicycle will walk))}

\item{nodes}{data.frame with columns "node", "X", and "Y".}

\item{crs}{character string representing the coordinate reference system.}

\item{directed}{logical value indicating whether the graph is directed (default is TRUE).}
}
\value{
A Graph object.
}
\description{
This function takes edge and node data.frames and a CRS string to create a Graph object.
}
\examples{
\dontrun{
edges <- data.frame(from = c("A", "A", "B", "C"),
                    to = c("B", "C", "C", "D"),
                    speed = c(10, 20, 40, 100), # km/h
                    length = c(1, 2, 2, 1),     # m
                    oneway = c("FT", "B", "N", "TF"))

nodes <- data.frame(node = c("A", "B", "C", "D"),
                    X = c(0, 1, 1, 2),
                    Y = c(0, 0, 1, 1))

crs <- "EPSG:4326"

graph <- makegraph(edges, nodes, crs, directed = TRUE)
print(graph)
}

}
